requirements.txt
python-dotenv
flask
flask_sqlalchemy
flask_migrate==2.5.3
flask_script
cryptography